backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
httpx>=0.27.0
websockets>=13.0,<15.0
google-genai>=1.9.0
google-adk>=0.1.0
pydantic>=2.9.0
